converted react-scripts to vite
